Refactored FilesResponse into entities, aggregated entities and response models #431
