+++ explorer/app/apis/azul/hca-dcp/common/entities.ts
@@ -1,6 +1,3 @@
-/**
- * Model of response returned from /index/files API endpoint.
- */
 export interface FilesEntity {
   contentDescription: string[];
   format: string;
@@ -14,24 +11,12 @@ export interface FilesEntityResponse {
   files: FilesEntity[];
 }
 
-/**
- * Model of response returned from /index/samples API endpoint.
- */
-export interface SamplesResponse {
-  donorOrganisms: {
-    disease: string[];
-    genusSpecies: string[];
-  }[];
-  projects: {
-    estimatedCellCount?: number;
-    projectTitle: string[];
-  }[];
-  protocols: {
-    libraryConstructionApproach: string[];
-  }[];
-  samples: {
-    id: string;
-    organ: string;
-    sampleEntityType: string;
-  }[];
+export interface SamplesEntity {
+  id: string;
+  organ: string;
+  sampleEntityType: string;
+}
+
+export interface SamplesEntityResponse {
+  samples: SamplesEntity[];
 }
