+++ explorer/app/apis/azul/hca-dcp/common/responses.ts
@@ -1,7 +1,23 @@
 import { AzulHit } from "../../common/entities";
-import { FilesEntityResponse } from "./entities";
-import { AggregatedProjectResponse } from "./aggregatedEntities";
+import { FilesEntityResponse, SamplesEntityResponse } from "./entities";
+import {
+  AggregatedDonorOrganismsResponse,
+  AggregatedProjectsResponse,
+  AggregatedProtocolsResponse,
+} from "./aggregatedEntities";
 
+/**
+ * Model of response returned from /index/files API endpoint.
+ */
 export type FilesResponse = AzulHit &
   FilesEntityResponse &
-  AggregatedProjectResponse;
+  AggregatedProjectsResponse;
+
+/**
+ * Model of response returned from /index/samples API endpoint.
+ */
+export type SamplesResponse = AzulHit &
+  SamplesEntityResponse &
+  AggregatedDonorOrganismsResponse &
+  AggregatedProjectsResponse &
+  AggregatedProtocolsResponse;
