+++ explorer/app/apis/azul/hca-dcp/common/transformers.ts
@@ -1,7 +1,6 @@
 import { concatStrings } from "../../../../utils/string";
 import { humanFileSize } from "../../../../utils/fileSize";
-import { FilesResponse } from "./responses";
-import { SamplesResponse } from "./entities";
+import { FilesResponse, SamplesResponse } from "./responses";
 
 //FilesResponse
